Minor fix in initial guess code
The initial guess was incorrectly assembled for cases where the background
current is nonzero, and this may also result in DJLES reporting incorrect
WNL coefficients r10/r01. However, since the initial guess is used to
initialize the iterative procedure, the error does not persist to the final
DJL solutions, which were unaffected and remain correct.